feat: Phase 5 - Task 6.4 完了 (Multi-Repoコマンド移行)
## 変更内容

### ✅ Task 6.4: Multi-Repoコマンドの移行（完了）

#### 1. multi-repo-list コマンドの移行
- `src/presentation/commands/multi-repo/list.ts` (61行)
- `src/commands/multi-repo-list.ts` (10行) - 薄いラッパー

#### 2. multi-repo-add-repo コマンドの移行
- `src/presentation/commands/multi-repo/add-repo.ts` (96行)
- `src/commands/multi-repo-add-repo.ts` (9行) - 薄いラッパー

#### 3. multi-repo-test コマンドの移行
- `src/presentation/commands/multi-repo/test.ts` (150行)
- `src/commands/multi-repo-test.ts` (13行) - 薄いラッパー

#### 4. multi-repo-init コマンドの移行
- `src/presentation/commands/multi-repo/init.ts` (159行)
- `src/commands/multi-repo-init.ts` (9行) - 薄いラッパー

#### 5. multi-repo-confluence-sync コマンドの移行
- `src/presentation/commands/multi-repo/confluence-sync.ts` (213行)
- `src/commands/multi-repo-confluence-sync.ts` (12行) - 薄いラッパー

#### 6. multi-repo-ci-status コマンドの移行
- `src/presentation/commands/multi-repo/ci-status.ts` (394行)
- `src/commands/multi-repo-ci-status.ts` (12行) - 薄いラッパー

## テスト結果

✅ **ビルド成功**
\`\`\`
npm run build
\`\`\`

✅ **全テスト成功**
\`\`\`
npm test
Test Suites: 925 passed, 5 skipped
\`\`\`

## 後方互換性

すべての既存のインポートパスは引き続き動作します。
